requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-environ==0.4.5
django-filter==2.4.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
idna==2.10
importlib-metadata==3.7.3
kombu==4.6.11
Markdown==3.3.4
prompt-toolkit==3.0.17
psycopg2==2.8.6
psycopg2-binary==2.8.5
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.1
